refactor(runtime): canonicalize Handle as a runtime primitive (#17 slice 1)
Move the universal correlation Handle(Uuid) newtype out of live/ — where it
read as a media-only type used by 6 av files — into runtime/, beside HandleRef
and CommandRequest where it belongs as a substrate primitive. Add ts-rs derive
so it crosses the wire as a bare uuid string (regression-guarded), making it
reusable in the envelope/result/event layers.

This is the foundation for reconciling the two handle models (task #17): Handle
is the bare correlation tag; HandleRef is Handle + routing metadata. Next slice
makes HandleRef.id a Handle so there is one uuid type referenced explicitly
everywhere, per Joel's directive.
